[fix] 옵션 리스트의 타입을 List<Long>으로 수정
[issue] #248

+++ BackEnd/FourEver/src/main/java/A4/FourEver/domain/car/application/CarService.java
@@ -9,6 +9,6 @@ import java.util.List;
 public interface CarService {
     CarTrimsSortedDTO getCarTrimsById(final Long id);
     CarExtraOptionNameDTO getCarExtraOptionNameById(final Long id);
-    CarReviewOverviewSortedListDTO getAllCarReviewOverviewList(final Long id, final List<Integer> extraOptionIds);
-    CarReviewOverviewSortedListDTO getPartialCarReviewOverviewList(final Long id, final Integer isPurchased, final List<Integer> extraOptionIds);
+    CarReviewOverviewSortedListDTO getAllCarReviewOverviewList(final Long id, final List<Long> extraOptionIds);
+    CarReviewOverviewSortedListDTO getPartialCarReviewOverviewList(final Long id, final Integer isPurchased, final List<Long> extraOptionIds);
 }
